requirements.txt
asyncio
websockets
pygame
requests
numpy
pyyaml